Default batch size to 1000 and remove extra query
Change the default batch size to 1000 results, anecdotally real-world
latency seems to have more of an impact to the overall speed of results.

When we are querying results without following, if the last result batch
was smaller than our specified result size, assume we have reached the
end and exit early to avoid an additional query that could have zero
results.